refactor(app): batch9 — extract useAppShellState orchestration hub, move AppShell to app-shell.tsx, shrink App.tsx to 18 lines
Split the AppShell orchestration core into composable sub-hooks so every
file lands under 300 lines:
- use-app-shell-state.tsx (268): orchestration hub combining the four clusters
- use-selection-controllers.ts (158): TDZ ring (selection/contextPanel/startSurface + pivots + pendingCreated)
- use-shell-chrome-state.ts (114): chrome/panels state + actions
- use-workspace-data-controllers.ts (200): forge/editor/read/commit/session/nav/quickSwitch/pendingQueue
- use-workspace-action-controllers.ts (240): workspace action handlers
- app-shell.tsx (242): AppShell — pulls useAppShellState, computes the 2 perf-critical header memos at the orchestration boundary, assembles child-prop bags
- app-shell-layout.tsx (133): pure presentational assembly shell

Verbatim move + threading only. Header memo deps and the pendingCreated
callback deps kept identical to the original (P1-A / doNotBreak red lines).
